tests/e2e-tests/requirements.txt
pytest-playwright
pytest-html
python-dotenv